% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{aaGabriel2022Direct}
\alias{aaGabriel2022Direct}
\title{Direct modeling of relative and absolute risks in register data: Mortality risk in sarcoidosis (2022) \emph{Annals of epidemiology}}
\description{
Gabriel, EE; Arkema, EV; Sachs, MC
}
\details{
Annals of epidemiology (2022). 1-4.:NA. \url{http://www.ncbi.nlm.nih.gov/pubmed/34775031}
}
\concept{methods}
