% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{aaRundquist2021Drug}
\alias{aaRundquist2021Drug}
\title{Drug survival of anti-TNF agents compared with vedolizumab as a second-line biological treatment in inflammatory bowel disease: results from nationwide Swedish registers (2021) \emph{ALIMENTARY PHARMACOLOGY & THERAPEUTICS}}
\description{
Rundquist, S; Sachs, MC; Eriksson, C; Olen, O; Montgomery, S; Halfvarson, J; SWIBREGStudyGrp
}
\details{
ALIMENTARY PHARMACOLOGY & THERAPEUTICS (2021). 471-483:53. \url{http://www.ncbi.nlm.nih.gov/pubmed/33340426}
}
\concept{applied}
