% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{abDehara2022Reproductive}
\alias{abDehara2022Reproductive}
\title{Reproductive and hormonal risk factors for sarcoidosis: a nested case-control study (2022) \emph{BMC pulmonary medicine}}
\description{
Dehara, M; Sachs, MC; Kullberg, S; Grunewald, J; Blomberg, A; Arkema, EV
}
\details{
BMC pulmonary medicine (2022). 43-.:1. \url{http://www.ncbi.nlm.nih.gov/pubmed/35073900}
}
\concept{applied}
