% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{acLudvigsson2021Adult-onset}
\alias{acLudvigsson2021Adult-onset}
\title{Adult-onset inflammatory bowel disease and rate of serious infections compared to the general population: a nationwide register-based cohort study 2002-2017 (2021) \emph{Scandinavian journal of gastroenterology}}
\description{
Ludvigsson, JF; Holmgren, J; Grip, O; Halfvarson, J; Askling, J; Sachs, MC; Olén, O
}
\details{
Scandinavian journal of gastroenterology (2021). 1152-1162.:10. \url{http://www.ncbi.nlm.nih.gov/pubmed/34369254}
}
\concept{applied}
