% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{adLochhead2021Association}
\alias{adLochhead2021Association}
\title{Association Between Statin Use and Inflammatory Bowel Diseases: Results from a Swedish, Nationwide, Population-based Case-control Study (2021) \emph{Journal of Crohn's & colitis}}
\description{
Lochhead, P; Khalili, H; Sachs, MC; Chan, AT; Olén, O; Ludvigsson, JF
}
\details{
Journal of Crohn's & colitis (2021). 757-765.:5. \url{http://www.ncbi.nlm.nih.gov/pubmed/33216873}
}
\concept{applied}
