% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{adLochhead2021Statin}
\alias{adLochhead2021Statin}
\title{Statin use and risk of inflammatory bowel diseases: authors' reply (2021) \emph{Journal of Crohn's & colitis}}
\description{
Lochhead, P; Khalili, H; Sachs, MC; Chan, AT; Ol<e9>n, O; Ludvigsson, JF
}
\details{
Journal of Crohn's & colitis (2021). NA:NA. \url{http://www.ncbi.nlm.nih.gov/pubmed/33474560}
}
\concept{applied}
