% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{aeBröms2021Capturing}
\alias{aeBröms2021Capturing}
\title{Capturing biologic treatment for IBD in the Swedish Prescribed Drug Register and the Swedish National Patient Register - a validation study (2021) \emph{Scandinavian journal of gastroenterology}}
\description{
Bröms, G; Söderling, J; Sachs, MC; Halfvarson, J; Myrelid, P; Ludvigsson, JF; Everhov, ÅH; Olén, O; SWIBREG study group
}
\details{
Scandinavian journal of gastroenterology (2021). 410-421.:4. \url{http://www.ncbi.nlm.nih.gov/pubmed/33632044}
}
\concept{applied}
