% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{afSaccon2021Distinct}
\alias{afSaccon2021Distinct}
\title{Distinct Metabolic Profile Associated with a Fatal Outcome in COVID-19 Patients during the Early Epidemic in Italy (2021) \emph{Microbiology spectrum}}
\description{
Saccon, E; Bandera, A; Sciumè, M; Mikaeloff, F; Lashari, AA; Aliberti, S; Sachs, MC; Billi, F; Blasi, F; Gabriel, EE; Costantino, G; De Roberto, P; Krishnan, S; Gori, A; Peyvandi, F; Scudeller, L; Canetta, C; Lorson, CL; Valenti, L; Singh, K; Baldini, L; Fracchiolla, NS; Neogi, U; COVID-19 Network Working Group,
}
\details{
Microbiology spectrum (2021). e0054921-.:2. \url{http://www.ncbi.nlm.nih.gov/pubmed/34468185}
}
\concept{applied}
