% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{afSachs2020Aim}
\alias{afSachs2020Aim}
\title{Aim for Clinical Utility, Not Just Predictive Accuracy (2020) \emph{Epidemiology (Cambridge, Mass.)}}
\description{
Sachs, MC; Sj<f6>lander, A; Gabriel, EE
}
\details{
Epidemiology (Cambridge, Mass.) (2020). 359-364:31. \url{http://www.ncbi.nlm.nih.gov/pubmed/32091429}
}
\concept{methods}
