% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{agAxelrad2020A}
\alias{agAxelrad2020A}
\title{A Novel Method for Quantifying Intestinal Inflammatory Burden in Inflammatory Bowel Disease Using Register Data (2020) \emph{Clinical epidemiology}}
\description{
Axelrad, JE; Sachs, MC; Ludvigsson, JF; Ol<e9>n, O; SWIBREGStudyGroup
}
\details{
Clinical epidemiology (2020). 1059-1072:12. \url{http://www.ncbi.nlm.nih.gov/pubmed/33116900}
}
\concept{methods}
