% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{ahLochhead2020Association}
\alias{ahLochhead2020Association}
\title{Association between statin use and inflammatory bowel diseases: Results from a Swedish, nationwide, population-based case-control study (2020) \emph{Journal of Crohn's & colitis}}
\description{
Lochhead, P; Khalili, H; Sachs, MC; Chan, AT; Ol<e9>n, O; Ludvigsson, JF
}
\details{
Journal of Crohn's & colitis (2020). NA:NA. \url{http://www.ncbi.nlm.nih.gov/pubmed/33216873}
}
\concept{applied}
