% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{aiErichsen2021Hepatobiliary}
\alias{aiErichsen2021Hepatobiliary}
\title{Hepatobiliary Cancer Risk in Patients with Inflammatory Bowel Disease: A Scandinavian Population-Based Cohort Study (2021) \emph{Cancer epidemiology, biomarkers & prevention : a publication of the American Association for Cancer Research, cosponsored by the American Society of Preventive Oncology}}
\description{
Erichsen, R; Olén, O; Sachs, MC; Pedersen, L; Halfvarson, J; Askling, J; Ekbom, A; Ludvigsson, JF; Sørensen, HT
}
\details{
Cancer epidemiology, biomarkers & prevention : a publication of the American Association for Cancer Research, cosponsored by the American Society of Preventive Oncology (2021). 886-894.:5. \url{http://www.ncbi.nlm.nih.gov/pubmed/33627380}
}
\concept{applied}
