% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{ajAxelrad2021Inflammatory}
\alias{ajAxelrad2021Inflammatory}
\title{Inflammatory bowel disease and risk of small bowel cancer: a binational population-based cohort study from Denmark and Sweden (2021) \emph{Gut}}
\description{
Axelrad, JE; Olén, O; Sachs, MC; Erichsen, R; Pedersen, L; Halfvarson, J; Askling, J; Ekbom, A; Sørensen, HT; Ludvigsson, JF
}
\details{
Gut (2021). 297-308.:2. \url{http://www.ncbi.nlm.nih.gov/pubmed/32474410}
}
\concept{applied}
