% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{ajGabriel2020Causal}
\alias{ajGabriel2020Causal}
\title{Causal Bounds for Outcome-Dependent Sampling in Observational Studies (2020) \emph{JOURNAL OF THE AMERICAN STATISTICAL ASSOCIATION}}
\description{
Gabriel, EE; Sachs, MC; Sjolander, A
}
\details{
JOURNAL OF THE AMERICAN STATISTICAL ASSOCIATION (2020). NA:NA. \url{NA}
}
\concept{methods}
