% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{akOl<U+653C><U+3E39>n2020Colorectal}
\alias{akOl<U+653C><U+3E39>n2020Colorectal}
\title{Colorectal cancer in Crohn's disease: a Scandinavian population-based cohort study (2020) \emph{The lancet. Gastroenterology & hepatology}}
\description{
Ol<e9>n, O; Erichsen, R; Sachs, MC; Pedersen, L; Halfvarson, J; Askling, J; Ekbom, A; S<f8>rensen, HT; Ludvigsson, JF
}
\details{
The lancet. Gastroenterology & hepatology (2020). 475-484:5. \url{http://www.ncbi.nlm.nih.gov/pubmed/32066530}
}
\concept{applied}
