% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{akRundquist2021Letter:}
\alias{akRundquist2021Letter:}
\title{Letter: vedolizumab or a second anti-TNF-no difference in efficacy for primary biologic failures with IBD. Authors' reply (2021) \emph{Alimentary pharmacology & therapeutics}}
\description{
Rundquist, S; Sachs, MC; Eriksson, C; Olén, O; Montgomery, S; Halfvarson, J
}
\details{
Alimentary pharmacology & therapeutics (2021). 1046-1047.:9. \url{http://www.ncbi.nlm.nih.gov/pubmed/33831234}
}
\concept{applied}
