% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{alChen2021No}
\alias{alChen2021No}
\title{No evidence of substantial underreporting of COVID-19 deaths in Taiwan during 2020 (2021) \emph{Journal of the Formosan Medical Association = Taiwan yi zhi}}
\description{
Chen, EY; Sachs, MC; Dickman, PW
}
\details{
Journal of the Formosan Medical Association = Taiwan yi zhi (2021). 1788-1789.:9. \url{http://www.ncbi.nlm.nih.gov/pubmed/33832837}
}
\concept{applied}
