% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{alOl<U+653C><U+3E39>n2020Colorectal}
\alias{alOl<U+653C><U+3E39>n2020Colorectal}
\title{Colorectal cancer in ulcerative colitis: a Scandinavian population-based cohort study (2020) \emph{Lancet (London, England)}}
\description{
Ol<e9>n, O; Erichsen, R; Sachs, MC; Pedersen, L; Halfvarson, J; Askling, J; Ekbom, A; S<f8>rensen, HT; Ludvigsson, JF
}
\details{
Lancet (London, England) (2020). 123-131:395. \url{http://www.ncbi.nlm.nih.gov/pubmed/31929014}
}
\concept{applied}
