% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{amGabriel2021Nonparametric}
\alias{amGabriel2021Nonparametric}
\title{Nonparametric Bounds for Causal Effects in Imperfect Randomized Experiments (2021) \emph{JOURNAL OF THE AMERICAN STATISTICAL ASSOCIATION}}
\description{
Gabriel, EE; Sjolander, A; Sachs, MC
}
\details{
JOURNAL OF THE AMERICAN STATISTICAL ASSOCIATION (2021). NA:NA. \url{NA}
}
\concept{methods}
