% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{amKalman2020Decrease}
\alias{amKalman2020Decrease}
\title{Decrease in primary but not in secondary abdominal surgery for Crohn's disease: nationwide cohort study, 1990-2014 (2020) \emph{BRITISH JOURNAL OF SURGERY}}
\description{
Kalman, TD; Everhov, AH; Nordenvall, C; Sachs, MC; Halfvarson, J; Ekbom, A; Ludvigsson, JE; Myrelid, P; Olen, O
}
\details{
BRITISH JOURNAL OF SURGERY (2020). 1529-1538:107. \url{http://www.ncbi.nlm.nih.gov/pubmed/32452553}
}
\concept{applied}
