% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{anGabriel2021ON}
\alias{anGabriel2021ON}
\title{ON THE USE OF NONPARAMETRIC BOUNDS FOR CAUSAL EFFECTS IN NULL RANDOMIZED TRIALS (2021) \emph{American journal of epidemiology}}
\description{
Gabriel, EE; Sachs, MC
}
\details{
American journal of epidemiology (2021). 2231-.:10. \url{http://www.ncbi.nlm.nih.gov/pubmed/34017983}
}
\concept{methods}
