% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{aoAxelrad2021Reply:}
\alias{aoAxelrad2021Reply:}
\title{Reply: Survival in Crohn's disease-associated small bowel adenocarcinoma (2021) \emph{Gut}}
\description{
Axelrad, JE; Olén, O; Sachs, MC; Erichsen, R; Pedersen, L; Halfvarson, J; Askling, J; Ekbom, A; Sørensen, HT; Ludvigsson, JF
}
\details{
Gut (2021). 998-.:5. \url{http://www.ncbi.nlm.nih.gov/pubmed/32873695}
}
\concept{applied}
