% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{aoKhalili2020Microscopic}
\alias{aoKhalili2020Microscopic}
\title{Microscopic Colitis and Risk of Inflammatory Bowel Disease in a Nationwide Cohort Study (2020) \emph{Gastroenterology}}
\description{
Khalili, H; Burke, KE; Roelstraete, B; Sachs, MC; Ol<e9>n, O; Ludvigsson, JF
}
\details{
Gastroenterology (2020). 1574-1583.e2:158. \url{http://www.ncbi.nlm.nih.gov/pubmed/31926169}
}
\concept{applied}
