% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{apLudvigsson2021Serious}
\alias{apLudvigsson2021Serious}
\title{Serious Infections in Pediatric Inflammatory Bowel Disease 2002-2017-A Nationwide Cohort Study (2021) \emph{The Journal of pediatrics}}
\description{
Ludvigsson, JF; Sachs, MC; Järås, J; Malmborg, P; Olén, O
}
\details{
The Journal of pediatrics (2021). 66-73.e1.:NA. \url{http://www.ncbi.nlm.nih.gov/pubmed/34216628}
}
\concept{applied}
