% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{apOl<U+653C><U+3E39>n2020Mortality}
\alias{apOl<U+653C><U+3E39>n2020Mortality}
\title{Mortality in adult-onset and elderly-onset IBD: a nationwide register-based cohort study 1964-2014 (2020) \emph{Gut}}
\description{
Ol<e9>n, O; Askling, J; Sachs, MC; Neovius, M; Smedby, KE; Ekbom, A; Ludvigsson, JF
}
\details{
Gut (2020). 453-461:69. \url{http://www.ncbi.nlm.nih.gov/pubmed/31092591}
}
\concept{applied}
