% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{arRundquist2021Editorial:}
\alias{arRundquist2021Editorial:}
\title{Editorial: anti-TNF agents against vedolizumab as a second-line treatment? Not surprising tie game-authors' reply (2021) \emph{Alimentary pharmacology & therapeutics}}
\description{
Rundquist, S; Sachs, MC; Eriksson, C; Ol<e9>n, O; Montgomery, S; Halfvarson, J
}
\details{
Alimentary pharmacology & therapeutics (2021). 642-643:53. \url{http://www.ncbi.nlm.nih.gov/pubmed/33566420}
}
\concept{applied}
