% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{asAxelrad2021Inflammatory}
\alias{asAxelrad2021Inflammatory}
\title{Inflammatory bowel disease and risk of small bowel cancer: a binational population-based cohort study from Denmark and Sweden (2021) \emph{Gut}}
\description{
Axelrad, JE; Ol<e9>n, O; Sachs, MC; Erichsen, R; Pedersen, L; Halfvarson, J; Askling, J; Ekbom, A; S<f8>rensen, HT; Ludvigsson, JF
}
\details{
Gut (2021). 297-308:70. \url{http://www.ncbi.nlm.nih.gov/pubmed/32474410}
}
\concept{applied}
