% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{asShrestha2020The}
\alias{asShrestha2020The}
\title{The use of ICD codes to identify IBD subtypes and phenotypes of the Montreal classification in the Swedish National Patient Register (2020) \emph{Scandinavian journal of gastroenterology}}
\description{
Shrestha, S; Ol<e9>n, O; Eriksson, C; Everhov, <c5>H; Myrelid, P; Visuri, I; Ludvigsson, JF; Schoultz, I; Montgomery, S; Sachs, MC; Halfvarson, J; Olsson, M; Hjortswang, H; Bengtsson, J; Strid, H; Andersson, M; J<e4>ghult, S; Eberhardson, M; Nordenvall, C; Bj<f6>rk, J; Fagerberg, UL; Rejler, M; Grip, O; Karling, P; Block, M; Angenete, E; Hellstr<f6>m, PM; Gustavsson, A; SWIBREGStudyGroup
}
\details{
Scandinavian journal of gastroenterology (2020). 430-435:55. \url{http://www.ncbi.nlm.nih.gov/pubmed/32370571}
}
\concept{methods}
