% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{auEverhov2020Work}
\alias{auEverhov2020Work}
\title{Work Loss in Relation to Pharmacological and Surgical Treatment for Crohn's Disease: A Population-Based Cohort Study (2020) \emph{Clinical epidemiology}}
\description{
Everhov, <c5>H; Sachs, MC; Ludvigsson, JF; Khalili, H; Askling, J; Neovius, M; Myrelid, P; Halfvarson, J; Nordenvall, C; S<f6>derling, J; Ol<e9>n, O; SWIBREGstudygroup
}
\details{
Clinical epidemiology (2020). 273-285:12. \url{http://www.ncbi.nlm.nih.gov/pubmed/32210631}
}
\concept{applied}
