% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{auMouratidou2020Adult}
\alias{auMouratidou2020Adult}
\title{Adult height in patients with childhood-onset inflammatory bowel disease: a nationwide population-based cohort study (2020) \emph{Alimentary pharmacology & therapeutics}}
\description{
Mouratidou, N; Malmborg, P; Sachs, MC; Askling, J; Ekbom, A; Neovius, M; Smedby, KE; S<e4>vendahl, L; Ludvigsson, JF; Ol<e9>n, O
}
\details{
Alimentary pharmacology & therapeutics (2020). 789-800:51. \url{http://www.ncbi.nlm.nih.gov/pubmed/32133656}
}
\concept{applied}
