% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{ayGabriel2020A}
\alias{ayGabriel2020A}
\title{A unified evaluation of differential vaccine efficacy (2020) \emph{BIOMETRICS}}
\description{
Gabriel, EE; Sachs, MC; Follmann, DA; Andersson, TML
}
\details{
BIOMETRICS (2020). 1053-1063:76. \url{http://www.ncbi.nlm.nih.gov/pubmed/31868914}
}
\concept{methods}
