% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{aySachs2019Ensemble}
\alias{aySachs2019Ensemble}
\title{Ensemble prediction of time-to-event outcomes with competing risks: a case-study of surgical complications in Crohn's disease (2019) \emph{JOURNAL OF THE ROYAL STATISTICAL SOCIETY SERIES C-APPLIED STATISTICS}}
\description{
Sachs, MC; Discacciati, A; Everhov, AH; Olen, O; Gabriel, EE
}
\details{
JOURNAL OF THE ROYAL STATISTICAL SOCIETY SERIES C-APPLIED STATISTICS (2019). 1431-1446:68. \url{NA}
}
\concept{methods}
