% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{baBj<U+663C><U+3E36>rkman2019From}
\alias{baBj<U+663C><U+3E36>rkman2019From}
\title{From high to low malaria transmission in Zanzibar-challenges and opportunities to achieve elimination (2019) \emph{BMC medicine}}
\description{
Bj<f6>rkman, A; Shakely, D; Ali, AS; Morris, U; Mkali, H; Abbas, AK; Al-Mafazy, AW; Haji, KA; Mcha, J; Omar, R; Cook, J; Elfving, K; Petzold, M; Sachs, MC; Aydin-Schmidt, B; Drakeley, C; Msellem, M; M<e5>rtensson, A
}
\details{
BMC medicine (2019). 14:17. \url{http://www.ncbi.nlm.nih.gov/pubmed/30665398}
}
\concept{applied}
