% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bcOl<U+653C><U+3E39>n2019Increased}
\alias{bcOl<U+653C><U+3E39>n2019Increased}
\title{Increased Mortality of Patients With Childhood-Onset Inflammatory Bowel Diseases, Compared With the General<U+00A0>Population (2019) \emph{Gastroenterology}}
\description{
Ol<e9>n, O; Askling, J; Sachs, MC; Frumento, P; Neovius, M; Smedby, KE; Ekbom, A; Malmborg, P; Ludvigsson, JF
}
\details{
Gastroenterology (2019). 614-622:156. \url{http://www.ncbi.nlm.nih.gov/pubmed/30342031}
}
\concept{applied}
