% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bdEverhov2020Inflammatory}
\alias{bdEverhov2020Inflammatory}
\title{Inflammatory bowel disease and pancreatic cancer: a Scandinavian register-based cohort study 1969-2017 (2020) \emph{Alimentary pharmacology & therapeutics}}
\description{
Everhov, <c5>H; Erichsen, R; Sachs, MC; Pedersen, L; Halfvarson, J; Askling, J; Ekbom, A; Ludvigsson, JF; S<f8>rensen, HT; Ol<e9>n, O
}
\details{
Alimentary pharmacology & therapeutics (2020). 143-154:52. \url{http://www.ncbi.nlm.nih.gov/pubmed/32412143}
}
\concept{applied}
