% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bdWeimers2019Inflammatory}
\alias{bdWeimers2019Inflammatory}
\title{Inflammatory Bowel Disease and Parkinson's Disease: A Nationwide Swedish Cohort Study (2019) \emph{Inflammatory bowel diseases}}
\description{
Weimers, P; Halfvarson, J; Sachs, MC; Saunders-Pullman, R; Ludvigsson, JF; Peter, I; Burisch, J; Ol<e9>n, O
}
\details{
Inflammatory bowel diseases (2019). 111-123:25. \url{http://www.ncbi.nlm.nih.gov/pubmed/29788069}
}
\concept{applied}
