% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bfCarithers2019The}
\alias{bfCarithers2019The}
\title{The Biospecimen Preanalytical Variables Program: A Multiassay Comparison of Effects of Delay to Fixation and Fixation Duration on Nucleic Acid Quality (2019) \emph{Archives of pathology & laboratory medicine}}
\description{
Carithers, LJ; Agarwal, R; Guan, P; Odeh, H; Sachs, MC; Engel, KB; Greytak, SR; Barcus, M; Soria, C; Lih, CJ; Williams, PM; Branton, PA; Sobin, L; Fombonne, B; Bocklage, T; Andry, C; Duffy, ER; Sica, G; Dhir, R; Jewell, S; Roche, N; Moore, HM; NationalCancerInstituteBiospecimenPreanalyticalVariablesResearchProgram
}
\details{
Archives of pathology & laboratory medicine (2019). 1106-1118:143. \url{http://www.ncbi.nlm.nih.gov/pubmed/30785788}
}
\concept{methods}
