% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bgKhalili2020Mortality}
\alias{bgKhalili2020Mortality}
\title{Mortality of Patients With Microscopic Colitis in Sweden (2020) \emph{Clinical gastroenterology and hepatology : the official clinical practice journal of the American Gastroenterological Association}}
\description{
Khalili, H; Bergman, D; Roelstraete, B; Burke, KE; Sachs, MC; Ol<e9>n, O; Ludvigsson, JF
}
\details{
Clinical gastroenterology and hepatology : the official clinical practice journal of the American Gastroenterological Association (2020). 2491-2499.e3:18. \url{http://www.ncbi.nlm.nih.gov/pubmed/31857243}
}
\concept{applied}
