% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bgSachs2019The}
\alias{bgSachs2019The}
\title{The Problem of Incomplete Validation in Reporting Prediction Models (2019) \emph{The American journal of gastroenterology}}
\description{
Sachs, MC
}
\details{
The American journal of gastroenterology (2019). 830-831:114. \url{http://www.ncbi.nlm.nih.gov/pubmed/31082841}
}
\concept{methods}
