% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bhAxelrad2020Reply:}
\alias{bhAxelrad2020Reply:}
\title{Reply: Survival in Crohn's disease-associated small bowel adenocarcinoma (2020) \emph{Gut}}
\description{
Axelrad, JE; Ol<e9>n, O; Sachs, MC; Erichsen, R; Pedersen, L; Halfvarson, J; Askling, J; Ekbom, A; S<f8>rensen, HT; Ludvigsson, JF
}
\details{
Gut (2020). NA:NA. \url{http://www.ncbi.nlm.nih.gov/pubmed/32873695}
}
\concept{applied}
