% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bhMorris2018A}
\alias{bhMorris2018A}
\title{A cluster randomised controlled trial of two rounds of mass drug administration in Zanzibar, a malaria pre-elimination setting-high coverage and safety, but no significant impact on transmission (2018) \emph{BMC medicine}}
\description{
Morris, U; Msellem, MI; Mkali, H; Islam, A; Aydin-Schmidt, B; Jovel, I; Shija, SJ; Khamis, M; Ali, SM; Hodzic, L; Magnusson, E; Poirot, E; Bennett, A; Sachs, MC; Tarning, J; M<e5>rtensson, A; Ali, AS; Bj<f6>rkman, A
}
\details{
BMC medicine (2018). 215:16. \url{http://www.ncbi.nlm.nih.gov/pubmed/30526588}
}
\concept{applied}
