% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Biostatistics-1}
\alias{Biostatistics-1}
\title{Optimizing and evaluating biomarker combinations as trial-level general surrogates (2019) \emph{Statistics in medicine}}
\description{
Gabriel, EE; Sachs, MC; Daniels, MJ; Halloran, ME
}
\details{
Statistics in medicine (2019). 1135-1146:38. \url{https://doi.org/10.1002/sim.7996}
}
