% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Biostatistics-10}
\alias{Biostatistics-10}
\title{Estimating mean annual 25-hydroxyvitamin D concentrations from single measurements: the Multi-Ethnic Study of Atherosclerosis (2013) \emph{AMERICAN JOURNAL OF CLINICAL NUTRITION}}
\description{
Sachs, MC; Shoben, A; Levin, GP; Robinson-Cohen, C; Hoofnagle, AN; Swords-Jenny, N; Ix, JH; Budoff, M; Lutsey, PL; Siscovick, DS; Kestenbaum, B; deBoer, IH
}
\details{
AMERICAN JOURNAL OF CLINICAL NUTRITION (2013). 1243-51:97. \url{https://doi.org/10.3945/ajcn.112.054502}
}
