% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Biostatistics-11}
\alias{Biostatistics-11}
\title{Partial summary measures of the predictiveness curve (2013) \emph{BIOMETRICAL JOURNAL}}
\description{
Sachs, MC; Zhou, XH
}
\details{
BIOMETRICAL JOURNAL (2013). 589-602:55. \url{https://doi.org/10.1002/bimj.201200146}
}
