% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Biostatistics-12}
\alias{Biostatistics-12}
\title{Ensemble prediction of time-to-event outcomes with competing risks: a case-study of surgical complications in Crohn's disease (NA) \emph{Journal of the Royal Statistical Society: Series C (Applied Statistics)}}
\description{
Sachs, Michael C.; Discacciati, Andrea; Everhov, Åsa H.; Olén, Ola; Gabriel, Erin E.
}
\details{
Journal of the Royal Statistical Society: Series C (Applied Statistics) (NA). NA:0. \url{https://rss.onlinelibrary.wiley.com/doi/abs/10.1111/rssc.12367}
}
