% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Biostatistics-2}
\alias{Biostatistics-2}
\title{The Problem of Incomplete Validation in Reporting Prediction Models (2019) \emph{The American journal of gastroenterology}}
\description{
Sachs, MC
}
\details{
The American journal of gastroenterology (2019). 830-831:114. \url{https://doi.org/10.14309/ajg.0000000000000188}
}
