% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Biostatistics-3}
\alias{Biostatistics-3}
\title{Evaluation and comparison of predictive individual-level general surrogates (2018) \emph{Biostatistics (Oxford, England)}}
\description{
Gabriel, EE; Sachs, MC; Halloran, ME
}
\details{
Biostatistics (Oxford, England) (2018). 307-324:19. \url{https://doi.org/10.1093/biostatistics/kxx037}
}
