% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Biostatistics-4}
\alias{Biostatistics-4}
\title{Measuring precision in bioassays: Rethinking assay validation (2018) \emph{Statistics in medicine}}
\description{
Fay, MP; Sachs, MC; Miura, K
}
\details{
Statistics in medicine (2018). 519-529:37. \url{https://doi.org/10.1002/sim.7528}
}
