% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Biostatistics-5}
\alias{Biostatistics-5}
\title{Comparing and combining biomarkers as principal surrogates for time-to-event clinical endpoints (2017) \emph{STATISTICS IN MEDICINE}}
\description{
Gabriel, EE; Sachs, MC; Gilbert, PB
}
\details{
STATISTICS IN MEDICINE (2017). 3440:36. \url{https://doi.org/10.1002/sim.7354}
}
