% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Biostatistics-6}
\alias{Biostatistics-6}
\title{plotROC: A Tool for Plotting ROC Curves (2017) \emph{Journal of statistical software}}
\description{
Sachs, MC
}
\details{
Journal of statistical software (2017). NA:79. \url{https://doi.org/10.18637/jss.v079.c02}
}
