% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Biostatistics-7}
\alias{Biostatistics-7}
\title{An Introduction to Principal Surrogate Evaluation with the pseval Package (2016) \emph{The R journal}}
\description{
Sachs, MC; Gabriel, EE
}
\details{
The R journal (2016). 277-292:8. \url{NA}
}
