% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Biostatistics-8}
\alias{Biostatistics-8}
\title{Issues in developing multivariable molecular signatures for guiding clinical care decisions (2016) \emph{Journal of biopharmaceutical statistics}}
\description{
Sachs, MC; McShane, LM
}
\details{
Journal of biopharmaceutical statistics (2016). 1098-1110:26. \url{https://doi.org/10.1080/10543406.2016.1226329}
}
