% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Biostatistics-9}
\alias{Biostatistics-9}
\title{Statistical principles for omics-based clinical trials (2015) \emph{Chinese clinical oncology}}
\description{
Sachs, MC
}
\details{
Chinese clinical oncology (2015). 29:4. \url{https://doi.org/10.3978/j.issn.2304-3865.2015.01.02}
}
