% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bjEverhov2018Incidence}
\alias{bjEverhov2018Incidence}
\title{Incidence and Treatment of Patients Diagnosed With<U+00A0>Inflammatory Bowel Diseases at 60 Years or Older<U+00A0>in<U+00A0>Sweden (2018) \emph{Gastroenterology}}
\description{
Everhov, <c5>H; Halfvarson, J; Myrelid, P; Sachs, MC; Nordenvall, C; S<f6>derling, J; Ekbom, A; Neovius, M; Ludvigsson, JF; Askling, J; Ol<e9>n, O
}
\details{
Gastroenterology (2018). 518-528.e15:154. \url{http://www.ncbi.nlm.nih.gov/pubmed/29102619}
}
\concept{applied}
