% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bjLiu2020Vagotomy}
\alias{bjLiu2020Vagotomy}
\title{Vagotomy and subsequent risk of inflammatory bowel disease: a nationwide register-based matched cohort study (2020) \emph{Alimentary pharmacology & therapeutics}}
\description{
Liu, B; Wanders, A; Wirdefeldt, K; Sj<f6>lander, A; Sachs, MC; Eberhardson, M; Ye, W; Ekbom, A; Ol<e9>n, O; Ludvigsson, JF
}
\details{
Alimentary pharmacology & therapeutics (2020). 1022-1030:51. \url{http://www.ncbi.nlm.nih.gov/pubmed/32319125}
}
\concept{applied}
