% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bkFay2018Measuring}
\alias{bkFay2018Measuring}
\title{Measuring precision in bioassays: Rethinking assay validation (2018) \emph{Statistics in medicine}}
\description{
Fay, MP; Sachs, MC; Miura, K
}
\details{
Statistics in medicine (2018). 519-529:37. \url{http://www.ncbi.nlm.nih.gov/pubmed/29052234}
}
\concept{methods}
