% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{blAmmerlaan2018Small}
\alias{blAmmerlaan2018Small}
\title{Small Nucleolar RNA Score: An Assay to Detect Formalin-Overfixed Tissue (2018) \emph{Biopreservation and biobanking}}
\description{
Ammerlaan, W; Trouet, J; Sachs, MC; Guan, P; Carithers, L; Lambert, P; Frasquilho, S; Antunes, L; Kofanova, O; Rohrer, D; Valley, DR; Blanski, A; Jewell, S; Moore, H; Betsou, F
}
\details{
Biopreservation and biobanking (2018). 467-476:16. \url{http://www.ncbi.nlm.nih.gov/pubmed/30234371}
}
\concept{methods}
