% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{blWeimers2019Association}
\alias{blWeimers2019Association}
\title{Association between inflammatory bowel disease and Parkinson's disease: seek and you shall find? (2019) \emph{Gut}}
\description{
Weimers, P; Halfvarson, J; Sachs, MC; Ludvigsson, JF; Peter, I; Ol<e9>n, O; Burisch, J
}
\details{
Gut (2019). 175-176:68. \url{http://www.ncbi.nlm.nih.gov/pubmed/30021791}
}
\concept{applied}
