% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bmEverhov2019Changes}
\alias{bmEverhov2019Changes}
\title{Changes in inflammatory bowel disease subtype during follow-up and over time in 44,302 patients (2019) \emph{Scandinavian journal of gastroenterology}}
\description{
Everhov, <c5>H; Sachs, MC; Malmborg, P; Nordenvall, C; Myrelid, P; Khalili, H; Elmberg, M; Ekbom, A; Askling, J; Jakobsson, G; Halfvarson, J; Ludvigsson, JF; Ol<e9>n, O
}
\details{
Scandinavian journal of gastroenterology (2019). 55-63:54. \url{http://www.ncbi.nlm.nih.gov/pubmed/30700170}
}
\concept{applied}
