% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bmOl<U+653C><U+3E39>n2017Childhood}
\alias{bmOl<U+653C><U+3E39>n2017Childhood}
\title{Childhood onset inflammatory bowel disease and risk of cancer: a Swedish nationwide cohort study 1964-2014 (2017) \emph{BMJ (Clinical research ed.)}}
\description{
Ol<e9>n, O; Askling, J; Sachs, MC; Frumento, P; Neovius, M; Smedby, KE; Ekbom, A; Malmborg, P; Ludvigsson, JF
}
\details{
BMJ (Clinical research ed.) (2017). j3951:358. \url{http://www.ncbi.nlm.nih.gov/pubmed/28931512}
}
\concept{applied}
