% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bnMalmborg2019Effects}
\alias{bnMalmborg2019Effects}
\title{Effects of Childhood-onset Inflammatory Bowel Disease on School Performance: A Nationwide Population-based Cohort Study Using Swedish Health and Educational Registers (2019) \emph{Inflammatory bowel diseases}}
\description{
Malmborg, P; Mouratidou, N; Sachs, MC; Hammar, U; Khalili, H; Neovius, M; Hjern, A; Smedby, KE; Ekbom, A; Askling, J; Ludvigsson, JF; Ol<e9>n, O
}
\details{
Inflammatory bowel diseases (2019). 1663-1673:25. \url{http://www.ncbi.nlm.nih.gov/pubmed/30916332}
}
\concept{applied}
