% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{boLucas2017Development}
\alias{boLucas2017Development}
\title{Development and validation of a Luminex assay for detection of a predictive biomarker for PROSTVAC-VF therapy (2017) \emph{PLOS ONE}}
\description{
Lucas, JL; Tacheny, EA; Ferris, A; Galusha, M; Srivastava, AK; Ganguly, A; Williams, PM; Sachs, MC; Thurin, M; Tricoli, JV; Ricker, W; Gildersleeve, JC
}
\details{
PLOS ONE (2017). e0182739:12. \url{http://www.ncbi.nlm.nih.gov/pubmed/28771597}
}
\concept{applied}
