% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bpLudvigsson2019Fracture}
\alias{bpLudvigsson2019Fracture}
\title{Fracture Risk in Patients With Inflammatory Bowel Disease: A Nationwide Population-Based Cohort Study From 1964 to 2014 (2019) \emph{The American journal of gastroenterology}}
\description{
Ludvigsson, JF; Mahl, M; Sachs, MC; Bj<f6>rk, J; Michaelsson, K; Ekbom, A; Askling, J; Backman, AS; Ol<e9>n, O
}
\details{
The American journal of gastroenterology (2019). 291-304:114. \url{http://www.ncbi.nlm.nih.gov/pubmed/30730858}
}
\concept{applied}
