% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bpSachs2017plotROC:}
\alias{bpSachs2017plotROC:}
\title{plotROC: A Tool for Plotting ROC Curves (2017) \emph{JOURNAL OF STATISTICAL SOFTWARE}}
\description{
Sachs, MC
}
\details{
JOURNAL OF STATISTICAL SOFTWARE (2017). NA:79. \url{http://www.ncbi.nlm.nih.gov/pubmed/30686944}
}
\concept{methods}
