% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{brAxelrad2019Gastrointestinal}
\alias{brAxelrad2019Gastrointestinal}
\title{Gastrointestinal Infection Increases Odds of Inflammatory Bowel Disease in a Nationwide Case-Control Study (2019) \emph{Clinical gastroenterology and hepatology : the official clinical practice journal of the American Gastroenterological Association}}
\description{
Axelrad, JE; Ol<e9>n, O; Askling, J; Lebwohl, B; Khalili, H; Sachs, MC; Ludvigsson, JF
}
\details{
Clinical gastroenterology and hepatology : the official clinical practice journal of the American Gastroenterological Association (2019). 1311-1322.e7:17. \url{http://www.ncbi.nlm.nih.gov/pubmed/30389589}
}
\concept{applied}
