% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{brSachs2016Issues}
\alias{brSachs2016Issues}
\title{Issues in developing multivariable molecular signatures for guiding clinical care decisions (2016) \emph{Journal of biopharmaceutical statistics}}
\description{
Sachs, MC; McShane, LM
}
\details{
Journal of biopharmaceutical statistics (2016). 1098-1110:26. \url{http://www.ncbi.nlm.nih.gov/pubmed/27561083}
}
\concept{methods}
