% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bsKorn2016Statistical}
\alias{bsKorn2016Statistical}
\title{Statistical controversies in clinical research: assessing pathologic complete response as a trial-level surrogate end point for early-stage breast cancer (2016) \emph{ANNALS OF ONCOLOGY}}
\description{
Korn, EL; Sachs, MC; McShane, LM
}
\details{
ANNALS OF ONCOLOGY (2016). 05-Oct:27. \url{http://www.ncbi.nlm.nih.gov/pubmed/26489443}
}
\concept{methods}
