% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{bwvanBallegooijen2014Association}
\alias{bwvanBallegooijen2014Association}
\title{Association of 25-Hydroxyvitamin D and Parathyroid Hormone With Incident Hypertension MESA (Multi-Ethnic Study of Atherosclerosis) (2014) \emph{JOURNAL OF THE AMERICAN COLLEGE OF CARDIOLOGY}}
\description{
vanBallegooijen, AJ; Kestenbaum, B; Sachs, MC; deBoer, IH; Siscovick, DS; Hoofnagle, AN; Ix, JH; Visser, M; Brouwer, IA
}
\details{
JOURNAL OF THE AMERICAN COLLEGE OF CARDIOLOGY (2014). 1214-1222:63. \url{http://www.ncbi.nlm.nih.gov/pubmed/24480627}
}
\concept{applied}
