% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{byGabriel2018Evaluation}
\alias{byGabriel2018Evaluation}
\title{Evaluation and comparison of predictive individual-level general surrogates (2018) \emph{Biostatistics (Oxford, England)}}
\description{
Gabriel, EE; Sachs, MC; Halloran, ME
}
\details{
Biostatistics (Oxford, England) (2018). 307-324:19. \url{http://www.ncbi.nlm.nih.gov/pubmed/28968890}
}
\concept{methods}
