% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Cardiology-10}
\alias{Cardiology-10}
\title{Total and Cause-Specific Mortality in the Cardiovascular Health Study (2009) \emph{JOURNALS OF GERONTOLOGY SERIES A-BIOLOGICAL SCIENCES AND MEDICAL SCIENCES}}
\description{
Newman, AB; Sachs, MC; Arnold, AM; Fried, LP; Kronmal, R; Cushman, M; Psaty, BM; Harris, TB; Robbins, JA; Burke, GL; Kuller, LH; Lumley, T
}
\details{
JOURNALS OF GERONTOLOGY SERIES A-BIOLOGICAL SCIENCES AND MEDICAL SCIENCES (2009). 1251-61:64. \url{https://doi.org/10.1093/gerona/glp127}
}
