% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Cardiology-2}
\alias{Cardiology-2}
\title{Fibroblast Growth Factor-23 and Cardiovascular Disease in the General Population The Multi-Ethnic Study of Atherosclerosis (2014) \emph{CIRCULATION-HEART FAILURE}}
\description{
Kestenbaum, B; Sachs, MC; Hoofnagle, AN; Siscovick, DS; Ix, JH; Robinson-Cohen, C; Lima, JAC; Polak, JF; Blondon, M; Ruzinski, J; Rock, D; deBoer, IH
}
\details{
CIRCULATION-HEART FAILURE (2014). 409-17:7. \url{https://doi.org/10.1161/CIRCHEARTFAILURE.113.000952}
}
