% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Cardiology-3}
\alias{Cardiology-3}
\title{Fibroblast Growth Factor-23 and Incident Atrial Fibrillation The Multi-Ethnic Study of Atherosclerosis (MESA) and the Cardiovascular Health Study (CHS) (2014) \emph{CIRCULATION}}
\description{
Mathew, JS; Sachs, MC; Katz, R; Patton, KK; Heckbert, SR; Hoofnagle, AN; Alonso, A; Chonchol, M; Deo, R; Ix, JH; Siscovick, DS; Kestenbaum, B; deBoer, IH
}
\details{
CIRCULATION (2014). 298-307:130. \url{https://doi.org/10.1161/CIRCULATIONAHA.113.005499}
}
