% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Cardiology-4}
\alias{Cardiology-4}
\title{Serum phosphate is associated with aortic valve calcification in the Multi-ethnic Study of Atherosclerosis (MESA) (2014) \emph{ATHEROSCLEROSIS}}
\description{
Linefsky, JP; O'Brien, KD; Sachs, M; Katz, R; Eng, J; Michos, ED; Budoff, MJ; deBoer, I; Kestenbaum, B
}
\details{
ATHEROSCLEROSIS (2014). 331-7:233. \url{https://doi.org/10.1016/j.atherosclerosis.2013.12.051}
}
