% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Cardiology-5}
\alias{Cardiology-5}
\title{25-Hydroxyvitamin D and Parathyroid Hormone Are Not Associated With Carotid Intima-Media Thickness or Plaque in the Multi-Ethnic Study of Atherosclerosis (2013) \emph{ARTERIOSCLEROSIS THROMBOSIS AND VASCULAR BIOLOGY}}
\description{
Blondon, M; Sachs, M; Hoofnagle, AN; Ix, JH; Michos, ED; Korcarz, C; Gepner, AD; Siscovick, DS; Kaufman, JD; Stein, JH; Kestenbaum, B; deBoer, IH
}
\details{
ARTERIOSCLEROSIS THROMBOSIS AND VASCULAR BIOLOGY (2013). 2639-45:33. \url{https://doi.org/10.1161/ATVBAHA.113.301781}
}
