% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Cardiology-6}
\alias{Cardiology-6}
\title{Parathyroid hormone and arterial dysfunction in the multi-ethnic study of atherosclerosis (2013) \emph{CLINICAL ENDOCRINOLOGY}}
\description{
Bosworth, C; Sachs, MC; Duprez, D; Hoofnagle, AN; Ix, JH; Jacobs, DR; Peralta, CA; Siscovick, DS; Kestenbaum, B; deBoer, IH
}
\details{
CLINICAL ENDOCRINOLOGY (2013). 429-36:79. \url{https://doi.org/10.1111/cen.12163}
}
