% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Cardiology-9}
\alias{Cardiology-9}
\title{Performance of Maximum Inspiratory Pressure Tests and Maximum Inspiratory Pressure Reference Equations for 4 Race/Ethnic Groups (2009) \emph{RESPIRATORY CARE}}
\description{
Multi-EthnicStudyAtherosclerosis; Sachs, MC; Enright, PL; Stukovsky, KDH; Jiang, R; Barr, RG
}
\details{
RESPIRATORY CARE (2009). 1321-8:54. \url{NA}
}
