% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{cbRobinson-Cohen2014Physical}
\alias{cbRobinson-Cohen2014Physical}
\title{Physical Activity and Change in Estimated GFR among Persons with CKD (2014) \emph{JOURNAL OF THE AMERICAN SOCIETY OF NEPHROLOGY}}
\description{
Robinson-Cohen, C; Littman, AJ; Duncan, GE; Weiss, NS; Sachs, MC; Ruzinski, J; Kundzins, J; Rock, D; deBoer, IH; Ikizler, TA; Himmelfarb, J; Kestenbaum, BR
}
\details{
JOURNAL OF THE AMERICAN SOCIETY OF NEPHROLOGY (2014). 399-406:25. \url{http://www.ncbi.nlm.nih.gov/pubmed/24335971}
}
\concept{applied}
