% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{cdBlondon201325-Hydroxyvitamin}
\alias{cdBlondon201325-Hydroxyvitamin}
\title{25-Hydroxyvitamin D and Parathyroid Hormone Are Not Associated With Carotid Intima-Media Thickness or Plaque in the Multi-Ethnic Study of Atherosclerosis (2013) \emph{ARTERIOSCLEROSIS THROMBOSIS AND VASCULAR BIOLOGY}}
\description{
Blondon, M; Sachs, M; Hoofnagle, AN; Ix, JH; Michos, ED; Korcarz, C; Gepner, AD; Siscovick, DS; Kaufman, JD; Stein, JH; Kestenbaum, B; deBoer, IH
}
\details{
ARTERIOSCLEROSIS THROMBOSIS AND VASCULAR BIOLOGY (2013). 2639-45:33. \url{http://www.ncbi.nlm.nih.gov/pubmed/23814117}
}
\concept{applied}
