% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{cdGabriel2017Comparing}
\alias{cdGabriel2017Comparing}
\title{Comparing and combining biomarkers as principal surrogates for time-to-event clinical endpoints (vol 34, pg 381, 2015) (2017) \emph{STATISTICS IN MEDICINE}}
\description{
Gabriel, EE; Sachs, MC; Gilbert, PB
}
\details{
STATISTICS IN MEDICINE (2017). 3440:36. \url{http://www.ncbi.nlm.nih.gov/pubmed/28776727}
}
\concept{methods}
