% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{chMehrotra2013No}
\alias{chMehrotra2013No}
\title{No independent association of serum phosphorus with risk for death or progression to end-stage renal disease in a large screen for chronic kidney disease (2013) \emph{KIDNEY INTERNATIONAL}}
\description{
Mehrotra, R; Peralta, CA; Chen, SC; Li, SY; Sachs, M; Shah, A; Norris, K; Saab, G; Whaley-Connell, A; Kestenbaum, B; McCullough, PA; KidneyEarlyEvaluationProgram
}
\details{
KIDNEY INTERNATIONAL (2013). 989-97:84. \url{http://www.ncbi.nlm.nih.gov/pubmed/23615501}
}
\concept{applied}
