% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{ciBosworth2013Parathyroid}
\alias{ciBosworth2013Parathyroid}
\title{Parathyroid hormone and arterial dysfunction in the multi-ethnic study of atherosclerosis (2013) \emph{CLINICAL ENDOCRINOLOGY}}
\description{
Bosworth, C; Sachs, MC; Duprez, D; Hoofnagle, AN; Ix, JH; Jacobs, DR; Peralta, CA; Siscovick, DS; Kestenbaum, B; deBoer, IH
}
\details{
CLINICAL ENDOCRINOLOGY (2013). 429-36:79. \url{http://www.ncbi.nlm.nih.gov/pubmed/23402353}
}
\concept{applied}
