% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{cjGabriel2015Comparing}
\alias{cjGabriel2015Comparing}
\title{Comparing and combining biomarkers as principle surrogates for time-to-event clinical endpoints (2015) \emph{STATISTICS IN MEDICINE}}
\description{
Gabriel, EE; Sachs, MC; Gilbert, PB
}
\details{
STATISTICS IN MEDICINE (2015). 381-95:34. \url{http://www.ncbi.nlm.nih.gov/pubmed/25352131}
}
\concept{methods}
