% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{clRobinson-Cohen2013Racial}
\alias{clRobinson-Cohen2013Racial}
\title{Racial Differences in the Association of Serum 25-Hydroxyvitamin D Concentration With Coronary Heart Disease Events (2013) \emph{JAMA-JOURNAL OF THE AMERICAN MEDICAL ASSOCIATION}}
\description{
Robinson-Cohen, C; Hoofnagle, AN; Ix, JH; Sachs, MC; Tracy, RP; Siscovick, DS; Kestenbaum, BR; deBoer, IH
}
\details{
JAMA-JOURNAL OF THE AMERICAN MEDICAL ASSOCIATION (2013). 179-88:310. \url{http://www.ncbi.nlm.nih.gov/pubmed/23839752}
}
\concept{applied}
