% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{clSachs2015Statistical}
\alias{clSachs2015Statistical}
\title{Statistical principles for omics-based clinical trials (2015) \emph{Chinese clinical oncology}}
\description{
Sachs, MC
}
\details{
Chinese clinical oncology (2015). 29:4. \url{http://www.ncbi.nlm.nih.gov/pubmed/26408296}
}
\concept{methods}
