% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{cnHernandez2010Pharmacological}
\alias{cnHernandez2010Pharmacological}
\title{Pharmacological treatment of Alzheimer's disease: effect of race and demographic variables (2010) \emph{Journal of Alzheimer's disease : JAD}}
\description{
Hernandez, S; McClendon, MJ; Zhou, XH; Sachs, M; Lerner, AJ
}
\details{
Journal of Alzheimer's disease : JAD (2010). 665-72:19. \url{http://www.ncbi.nlm.nih.gov/pubmed/20110610}
}
\concept{applied}
