% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{cnRingman2014Clinical}
\alias{cnRingman2014Clinical}
\title{Clinical Predictors of Severe Cerebral Amyloid Angiopathy and Influence of APOE Genotype in Persons With Pathologically Verified Alzheimer Disease (2014) \emph{JAMA NEUROLOGY}}
\description{
Ringman, JM; Sachs, MC; Zhou, Y; Monsell, SE; Saver, JL; Vinters, HV
}
\details{
JAMA NEUROLOGY (2014). 878-83:71. \url{http://www.ncbi.nlm.nih.gov/pubmed/24797962}
}
\concept{applied}
