% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{cpSachs2009Performance}
\alias{cpSachs2009Performance}
\title{Performance of Maximum Inspiratory Pressure Tests and Maximum Inspiratory Pressure Reference Equations for 4 Race/Ethnic Groups (2009) \emph{RESPIRATORY CARE}}
\description{
Sachs, MC; Enright, PL; Stukovsky, KDH; Jiang, R; Barr, RG; Multi-EthnicStudyAtherosclerosis
}
\details{
RESPIRATORY CARE (2009). 1321-8:54. \url{http://www.ncbi.nlm.nih.gov/pubmed/19796411}
}
\concept{applied}
