% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{cqMathew2014Fibroblast}
\alias{cqMathew2014Fibroblast}
\title{Fibroblast Growth Factor-23 and Incident Atrial Fibrillation The Multi-Ethnic Study of Atherosclerosis (MESA) and the Cardiovascular Health Study (CHS) (2014) \emph{CIRCULATION}}
\description{
Mathew, JS; Sachs, MC; Katz, R; Patton, KK; Heckbert, SR; Hoofnagle, AN; Alonso, A; Chonchol, M; Deo, R; Ix, JH; Siscovick, DS; Kestenbaum, B; deBoer, IH
}
\details{
CIRCULATION (2014). 298-307:130. \url{http://www.ncbi.nlm.nih.gov/pubmed/24920722}
}
\concept{applied}
