% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{cqNewman2009Total}
\alias{cqNewman2009Total}
\title{Total and Cause-Specific Mortality in the Cardiovascular Health Study (2009) \emph{JOURNALS OF GERONTOLOGY SERIES A-BIOLOGICAL SCIENCES AND MEDICAL SCIENCES}}
\description{
Newman, AB; Sachs, MC; Arnold, AM; Fried, LP; Kronmal, R; Cushman, M; Psaty, BM; Harris, TB; Robbins, JA; Burke, GL; Kuller, LH; Lumley, T
}
\details{
JOURNALS OF GERONTOLOGY SERIES A-BIOLOGICAL SCIENCES AND MEDICAL SCIENCES (2009). 1251-61:64. \url{http://www.ncbi.nlm.nih.gov/pubmed/19723772}
}
\concept{applied}
