% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{csLinefsky2014Serum}
\alias{csLinefsky2014Serum}
\title{Serum phosphate is associated with aortic valve calcification in the Multi-ethnic Study of Atherosclerosis (MESA) (2014) \emph{ATHEROSCLEROSIS}}
\description{
Linefsky, JP; O'Brien, KD; Sachs, M; Katz, R; Eng, J; Michos, ED; Budoff, MJ; deBoer, I; Kestenbaum, B
}
\details{
ATHEROSCLEROSIS (2014). 331-337:233. \url{http://www.ncbi.nlm.nih.gov/pubmed/24530958}
}
\concept{applied}
