% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{cuSachs2013Circulating}
\alias{cuSachs2013Circulating}
\title{Circulating Vitamin D Metabolites and Subclinical Atherosclerosis in Type 1 Diabetes (2013) \emph{DIABETES CARE}}
\description{
Sachs, MC; Brunzell, JD; Cleary, PA; Hoofnagle, AN; Lachin, JM; Molitch, ME; Steffes, MW; Zinman, B; deBoer, IH; DiabetControlComplicationsTrial
}
\details{
DIABETES CARE (2013). 23-Sep:36. \url{http://www.ncbi.nlm.nih.gov/pubmed/23530012}
}
\concept{applied}
