% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{cwAfkarian2013Kidney}
\alias{cwAfkarian2013Kidney}
\title{Kidney Disease and Increased Mortality Risk in Type 2 Diabetes (2013) \emph{JOURNAL OF THE AMERICAN SOCIETY OF NEPHROLOGY}}
\description{
Afkarian, M; Sachs, MC; Kestenbaum, B; Hirsch, IB; Tuttle, KR; Hinnmelfarb, J; deBoer, IH
}
\details{
JOURNAL OF THE AMERICAN SOCIETY OF NEPHROLOGY (2013). 302-8:24. \url{http://www.ncbi.nlm.nih.gov/pubmed/23362314}
}
\concept{applied}
