% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{czdeBoer2013Paricalcitol}
\alias{czdeBoer2013Paricalcitol}
\title{Paricalcitol does not improve glucose metabolism in patients with stage 3-4 chronic kidney disease (2013) \emph{KIDNEY INTERNATIONAL}}
\description{
deBoer, IH; Sachs, M; Hoofnagle, AN; Utzschneider, KM; Kahn, SE; Kestenbaum, B; Himmelfarb, J
}
\details{
KIDNEY INTERNATIONAL (2013). 323-30:83. \url{http://www.ncbi.nlm.nih.gov/pubmed/22913981}
}
\concept{applied}
