% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{daSachs2013Partial}
\alias{daSachs2013Partial}
\title{Partial summary measures of the predictiveness curve (2013) \emph{BIOMETRICAL JOURNAL}}
\description{
Sachs, MC; Zhou, XH
}
\details{
BIOMETRICAL JOURNAL (2013). 589-602:55. \url{http://www.ncbi.nlm.nih.gov/pubmed/23508801}
}
\concept{methods}
