% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{dcdeBoer2012Circulating}
\alias{dcdeBoer2012Circulating}
\title{Circulating Vitamin D Metabolites and Kidney Disease in Type 1 Diabetes (2012) \emph{JOURNAL OF CLINICAL ENDOCRINOLOGY & METABOLISM}}
\description{
deBoer, IH; Sachs, MC; Cleary, PA; Hoofnagle, AN; Lachin, JM; Molitch, ME; Steffes, MW; Sun, WJ; Zinman, B; Brunzell, JD; DiabetControlComplicationTrial
}
\details{
JOURNAL OF CLINICAL ENDOCRINOLOGY & METABOLISM (2012). Aug-80:97. \url{http://www.ncbi.nlm.nih.gov/pubmed/22990096}
}
\concept{applied}
