% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{deNewman2009Long-Term}
\alias{deNewman2009Long-Term}
\title{Long-Term Function in an Older Cohort-The Cardiovascular Health Study All Stars Study (2009) \emph{JOURNAL OF THE AMERICAN GERIATRICS SOCIETY}}
\description{
Newman, AB; Arnold, AM; Sachs, MC; Ives, DG; Cushman, M; Strotmeyer, ES; Ding, JZ; Kritchevsky, SB; Chaves, PHM; Fried, LP; Robbins, J
}
\details{
JOURNAL OF THE AMERICAN GERIATRICS SOCIETY (2009). 432-40:57. \url{http://www.ncbi.nlm.nih.gov/pubmed/19187412}
}
\concept{applied}
