% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Gastroenterology-10}
\alias{Gastroenterology-10}
\title{Childhood onset inflammatory bowel disease and risk of cancer: a Swedish nationwide cohort study 1964-2014 (2017) \emph{BMJ (Clinical research ed.)}}
\description{
Olén, O; Askling, J; Sachs, MC; Frumento, P; Neovius, M; Smedby, KE; Ekbom, A; Malmborg, P; Ludvigsson, JF
}
\details{
BMJ (Clinical research ed.) (2017). j3951:358. \url{https://doi.org/10.1136/bmj.j3951}
}
