% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Gastroenterology-2}
\alias{Gastroenterology-2}
\title{Changes in inflammatory bowel disease subtype during follow-up and over time in 44,302 patients (2019) \emph{Scandinavian journal of gastroenterology}}
\description{
Everhov, ÅH; Sachs, MC; Malmborg, P; Nordenvall, C; Myrelid, P; Khalili, H; Elmberg, M; Ekbom, A; Askling, J; Jakobsson, G; Halfvarson, J; Ludvigsson, JF; Olén, O
}
\details{
Scandinavian journal of gastroenterology (2019). 55-63:54. \url{https://doi.org/10.1080/00365521.2018.1564361}
}
