% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Gastroenterology-3}
\alias{Gastroenterology-3}
\title{Effects of Childhood-onset Inflammatory Bowel Disease on School Performance: A Nationwide Population-based Cohort Study Using Swedish Health and Educational Registers (2019) \emph{Inflammatory bowel diseases}}
\description{
Malmborg, P; Mouratidou, N; Sachs, MC; Hammar, U; Khalili, H; Neovius, M; Hjern, A; Smedby, KE; Ekbom, A; Askling, J; Ludvigsson, JF; Olén, O
}
\details{
Inflammatory bowel diseases (2019). NA:NA. \url{https://doi.org/10.1093/ibd/izz040}
}
