% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Gastroenterology-4}
\alias{Gastroenterology-4}
\title{Fracture Risk in Patients With Inflammatory Bowel Disease: A Nationwide Population-Based Cohort Study From 1964 to 2014 (2019) \emph{The American journal of gastroenterology}}
\description{
Ludvigsson, JF; Mahl, M; Sachs, MC; Björk, J; Michaelsson, K; Ekbom, A; Askling, J; Backman, AS; Olén, O
}
\details{
The American journal of gastroenterology (2019). 291-304:114. \url{https://doi.org/10.14309/ajg.0000000000000062}
}
