% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Gastroenterology-5}
\alias{Gastroenterology-5}
\title{Gastrointestinal Infection Increases Odds of Inflammatory Bowel Disease in a Nationwide Case-Control Study (2019) \emph{Clinical gastroenterology and hepatology : the official clinical practice journal of the American Gastroenterological Association}}
\description{
Axelrad, JE; Olén, O; Askling, J; Lebwohl, B; Khalili, H; Sachs, MC; Ludvigsson, JF
}
\details{
Clinical gastroenterology and hepatology : the official clinical practice journal of the American Gastroenterological Association (2019). 1311-1322.e7:17. \url{https://doi.org/10.1016/j.cgh.2018.09.034}
}
