% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Gastroenterology-6}
\alias{Gastroenterology-6}
\title{Increased Mortality of Patients With Childhood-Onset Inflammatory Bowel Diseases, Compared With the General Population (2019) \emph{Gastroenterology}}
\description{
Olén, O; Askling, J; Sachs, MC; Frumento, P; Neovius, M; Smedby, KE; Ekbom, A; Malmborg, P; Ludvigsson, JF
}
\details{
Gastroenterology (2019). 614-622:156. \url{https://doi.org/10.1053/j.gastro.2018.10.028}
}
