% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Gastroenterology-7}
\alias{Gastroenterology-7}
\title{Inflammatory Bowel Disease and Parkinson's Disease: A Nationwide Swedish Cohort Study (2019) \emph{Inflammatory bowel diseases}}
\description{
Weimers, P; Halfvarson, J; Sachs, MC; Saunders-Pullman, R; Ludvigsson, JF; Peter, I; Burisch, J; Olén, O
}
\details{
Inflammatory bowel diseases (2019). 111-123:25. \url{https://doi.org/10.1093/ibd/izy190}
}
