% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Gastroenterology-8}
\alias{Gastroenterology-8}
\title{Mortality in adult-onset and elderly-onset IBD: a nationwide register-based cohort study 1964-2014 (2019) \emph{Gut}}
\description{
Olén, O; Askling, J; Sachs, MC; Neovius, M; Smedby, KE; Ekbom, A; Ludvigsson, JF
}
\details{
Gut (2019). NA:NA. \url{https://doi.org/10.1136/gutjnl-2018-317572}
}
