% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Gastroenterology-9}
\alias{Gastroenterology-9}
\title{Incidence and Treatment of Patients Diagnosed With Inflammatory Bowel Diseases at 60 Years or Older in Sweden (2018) \emph{Gastroenterology}}
\description{
Everhov, ÅH; Halfvarson, J; Myrelid, P; Sachs, MC; Nordenvall, C; Söderling, J; Ekbom, A; Neovius, M; Ludvigsson, JF; Askling, J; Olén, O
}
\details{
Gastroenterology (2018). 518-528.e15:154. \url{https://doi.org/10.1053/j.gastro.2017.10.034}
}
