% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Infectious-disease-1}
\alias{Infectious-disease-1}
\title{From high to low malaria transmission in Zanzibar-challenges and opportunities to achieve elimination (2019) \emph{BMC medicine}}
\description{
Björkman, A; Shakely, D; Ali, AS; Morris, U; Mkali, H; Abbas, AK; Al-Mafazy, AW; Haji, KA; Mcha, J; Omar, R; Cook, J; Elfving, K; Petzold, M; Sachs, MC; Aydin-Schmidt, B; Drakeley, C; Msellem, M; Mårtensson, A
}
\details{
BMC medicine (2019). 14:17. \url{https://doi.org/10.1186/s12916-018-1243-z}
}
