% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Infectious-disease-2}
\alias{Infectious-disease-2}
\title{A cluster randomised controlled trial of two rounds of mass drug administration in Zanzibar, a malaria pre-elimination setting-high coverage and safety, but no significant impact on transmission (2018) \emph{BMC medicine}}
\description{
Morris, U; Msellem, MI; Mkali, H; Islam, A; Aydin-Schmidt, B; Jovel, I; Shija, SJ; Khamis, M; Ali, SM; Hodzic, L; Magnusson, E; Poirot, E; Bennett, A; Sachs, MC; Tarning, J; Mårtensson, A; Ali, AS; Björkman, A
}
\details{
BMC medicine (2018). 215:16. \url{https://doi.org/10.1186/s12916-018-1202-8}
}
