% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Lab-medicine-1}
\alias{Lab-medicine-1}
\title{The Biospecimen Preanalytical Variables Program: A Multiassay Comparison of Effects of Delay to Fixation and Fixation Duration on Nucleic Acid Quality (2019) \emph{Archives of pathology & laboratory medicine}}
\description{
NationalCancerInstituteBiospecimenPreanalyticalVariablesResearchProgram; Carithers, LJ; Agarwal, R; Guan, P; Odeh, H; Sachs, MC; Engel, KB; Greytak, SR; Barcus, M; Soria, C; Lih, CJ; Williams, PM; Branton, PA; Sobin, L; Fombonne, B; Bocklage, T; Andry, C; Duffy, ER; Sica, G; Dhir, R; Jewell, S; Roche, N; Moore, HM
}
\details{
Archives of pathology & laboratory medicine (2019). NA:NA. \url{https://doi.org/10.5858/arpa.2018-0172-OA}
}
