% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Lab-medicine-2}
\alias{Lab-medicine-2}
\title{Small Nucleolar RNA Score: An Assay to Detect Formalin-Overfixed Tissue (2018) \emph{Biopreservation and biobanking}}
\description{
Ammerlaan, W; Trouet, J; Sachs, MC; Guan, P; Carithers, L; Lambert, P; Frasquilho, S; Antunes, L; Kofanova, O; Rohrer, D; Valley, DR; Blanski, A; Jewell, S; Moore, H; Betsou, F
}
\details{
Biopreservation and biobanking (2018). NA:NA. \url{https://doi.org/10.1089/bio.2018.0042}
}
