% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Lab-medicine-4}
\alias{Lab-medicine-4}
\title{Robustness of Next Generation Sequencing on Older Formalin-Fixed Paraffin-Embedded Tissue (2015) \emph{PLOS ONE}}
\description{
Carrick, DM; Mehaffey, MG; Sachs, MC; Altekruse, S; Camalier, C; Chuaqui, R; Cozen, W; Das, B; Hernandez, BY; Lih, CJ; Lynch, CF; Makhlouf, H; McGregor, P; McShane, LM; Rohan, JP; Walsh, WD; Williams, PM; Gillanders, EM; Mechanic, LE; Schully, SD
}
\details{
PLOS ONE (2015). e0127353:10. \url{https://doi.org/10.1371/journal.pone.0127353}
}
