% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Nephrology-1}
\alias{Nephrology-1}
\title{Estimated GFR and Circulating 24,25-Dihydroxyvitamin D-3 Concentration: A Participant-Level Analysis of 5 Cohort Studies and Clinical Trials (2014) \emph{AMERICAN JOURNAL OF KIDNEY DISEASES}}
\description{
deBoer, IH; Sachs, MC; Chonchol, M; Himmelfarb, J; Hoofnagle, AN; Ix, JH; Kremsdorf, RA; Lin, YS; Mehrotra, R; Robinson-Cohen, C; Siscovick, DS; Steffes, MW; Thummel, KE; Tracy, RP; Wang, ZC; Kestenbaum, B
}
\details{
AMERICAN JOURNAL OF KIDNEY DISEASES (2014). 187-97:64. \url{https://doi.org/10.1053/j.ajkd.2014.02.015}
}
