% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Nephrology-3}
\alias{Nephrology-3}
\title{Circulating Vitamin D Metabolites and Subclinical Atherosclerosis in Type 1 Diabetes (2013) \emph{DIABETES CARE}}
\description{
DiabetControlComplicationsTrial; Sachs, MC; Brunzell, JD; Cleary, PA; Hoofnagle, AN; Lachin, JM; Molitch, ME; Steffes, MW; Zinman, B; deBoer, IH
}
\details{
DIABETES CARE (2013). 2423-9:36. \url{https://doi.org/10.2337/dc12-2020}
}
