% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Nephrology-4}
\alias{Nephrology-4}
\title{Kidney Disease and Increased Mortality Risk in Type 2 Diabetes (2013) \emph{JOURNAL OF THE AMERICAN SOCIETY OF NEPHROLOGY}}
\description{
Afkarian, M; Sachs, MC; Kestenbaum, B; Hirsch, IB; Tuttle, KR; Hinnmelfarb, J; deBoer, IH
}
\details{
JOURNAL OF THE AMERICAN SOCIETY OF NEPHROLOGY (2013). 302-8:24. \url{https://doi.org/10.1681/ASN.2012070718}
}
