% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Nephrology-5}
\alias{Nephrology-5}
\title{No independent association of serum phosphorus with risk for death or progression to end-stage renal disease in a large screen for chronic kidney disease (2013) \emph{KIDNEY INTERNATIONAL}}
\description{
KidneyEarlyEvaluationProgram; Mehrotra, R; Peralta, CA; Chen, SC; Li, SY; Sachs, M; Shah, A; Norris, K; Saab, G; Whaley-Connell, A; Kestenbaum, B; McCullough, PA
}
\details{
KIDNEY INTERNATIONAL (2013). 989-97:84. \url{https://doi.org/10.1038/ki.2013.145}
}
