% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Nephrology-6}
\alias{Nephrology-6}
\title{Paricalcitol does not improve glucose metabolism in patients with stage 3-4 chronic kidney disease (2013) \emph{KIDNEY INTERNATIONAL}}
\description{
deBoer, IH; Sachs, M; Hoofnagle, AN; Utzschneider, KM; Kahn, SE; Kestenbaum, B; Himmelfarb, J
}
\details{
KIDNEY INTERNATIONAL (2013). 323-30:83. \url{https://doi.org/10.1038/ki.2012.311}
}
