% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Nephrology-7}
\alias{Nephrology-7}
\title{Circulating Vitamin D Metabolites and Kidney Disease in Type 1 Diabetes (2012) \emph{JOURNAL OF CLINICAL ENDOCRINOLOGY & METABOLISM}}
\description{
DiabetControlComplicationTrial; deBoer, IH; Sachs, MC; Cleary, PA; Hoofnagle, AN; Lachin, JM; Molitch, ME; Steffes, MW; Sun, WJ; Zinman, B; Brunzell, JD
}
\details{
JOURNAL OF CLINICAL ENDOCRINOLOGY & METABOLISM (2012). 4780-8:97. \url{https://doi.org/10.1210/jc.2012-2852}
}
