% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Neurology-1}
\alias{Neurology-1}
\title{Clinical Predictors of Severe Cerebral Amyloid Angiopathy and Influence of APOE Genotype in Persons With Pathologically Verified Alzheimer Disease (2014) \emph{JAMA NEUROLOGY}}
\description{
Ringman, JM; Sachs, MC; Zhou, Y; Monsell, SE; Saver, JL; Vinters, HV
}
\details{
JAMA NEUROLOGY (2014). 878-83:71. \url{https://doi.org/10.1001/jamaneurol.2014.681}
}
