% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publications.R
\name{Oncology-1}
\alias{Oncology-1}
\title{Statistical controversies in clinical research: assessing pathologic complete response as a trial-level surrogate end point for early-stage breast cancer (2016) \emph{ANNALS OF ONCOLOGY}}
\description{
Korn, EL; Sachs, MC; McShane, LM
}
\details{
ANNALS OF ONCOLOGY (2016). 10-5:27. \url{https://doi.org/10.1093/annonc/mdv507}
}
